fix bugs

+++ src/main/java/com/snapp/price/snappprice/model/Location.java
@@ -3,6 +3,6 @@ package com.snapp.price.snappprice.model;
 import lombok.Data;
 @Data
 public class Location {
-    private final long x;
-    private final long y;
+    private final Double x;
+    private final Double y;
 }
